@@ -82,31 +77,32 @@
     "timers": "^0.1.1",
     "tippy.js": "^5.2.1",
     "tslib": "^2.2.0",
-    "typescript": "^4.4.4",
+    "typescript": "~4.6",
     "ua-parser-js": "^0.7.28",
-    "uuid": "^3.4.0",
+    "uuid": "^7.0.3",
+    "webpack": "^5.72.1",
     "wnumb": "^1.2.0",
     "xml2js": "^0.4.23",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular/language-service": "^13.0.3",
-    "@ngrx/schematics": "^10.1.2",
+    "@ngrx/schematics": "^13.2.0",
     "@types/jasmine": "^3.6.11",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "^12.20.12",
-    "codelyzer": "^6.0.2",
-    "jasmine-core": "~3.6.0",
-    "jasmine-marbles": "^0.6.0",
+    "codelyzer": "^0.0.28",
+    "jasmine-core": "~3.8.0",
+    "jasmine-marbles": "^0.9.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.4",
+    "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "nyc": "^15.1.0",
     "protractor": "~7.0.0",
-    "puppeteer": "^2.1.1",
+    "puppeteer": "^14.1.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^19.0.0",
     "ts-node": "^8.10.2",
