@@ -1,9 +1,9 @@
 import {
-  Component, OnInit, Input, ViewChild, ViewEncapsulation, OnDestroy
+  Component, OnInit, Input, ViewChild, ViewEncapsulation, OnDestroy, AfterContentInit
 } from '@angular/core';
 
 import { combineLatest, Observable } from 'rxjs';
-import { tap, withLatestFrom, filter, map, delay, debounceTime } from 'rxjs/operators';
+import { tap, withLatestFrom, filter, map, delay, debounceTime, first } from 'rxjs/operators';
 import { SubSink } from 'subsink';
 
 import { Store } from '@ngrx/store';
@@ -24,7 +24,7 @@ import { QueuedHyp3Job, SarviewsEvent } from '@models';
   styleUrls: ['./scenes-list.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class ScenesListComponent implements OnInit, OnDestroy {
+export class ScenesListComponent implements OnInit, OnDestroy, AfterContentInit {
   @ViewChild(CdkVirtualScrollViewport, { static: true }) scroll: CdkVirtualScrollViewport;
   @Input() resize$: Observable<void>;
 
@@ -261,6 +261,32 @@ export class ScenesListComponent implements OnInit, OnDestroy {
       ).subscribe(
         _ => this.scroll.scrollToOffset(0)
       )
+    );
+  }
+
+  ngAfterContentInit() {
+    this.subs.add(this.eventMonitoringService.filteredSarviewsEvents$().pipe(
+  filter(loaded => !!loaded),
+  withLatestFrom(this.store$.select(scenesStore.getSelectedSarviewsEvent)),
+    map(([events, selected]) => ({selectedEvent: selected, events})),
+        delay(400),
+        filter(selected => !!selected.selectedEvent),
+        tap(selected => {
+          this.mapService.zoomToEvent(selected.selectedEvent);
+          this.selectedEvent = selected.selectedEvent.event_id;
+        }),
+        first(),
+        map(selected => {
+          const sceneIdx = selected.events.findIndex(event => event.event_id === selected.selectedEvent.event_id);
+          return Math.max(0, sceneIdx - 1);
+        })
+      ).subscribe(
+        idx => {
+          if (!this.selectedFromList) {
+            this.scrollTo(idx);
+          }
+        }
+    )
     );
   }
 
