@@ -3,24 +3,27 @@ import { Component, OnInit, ViewChild, ElementRef, OnDestroy } from '@angular/co
 import { combineLatest } from 'rxjs';
 import { map, tap, filter, } from 'rxjs/operators';
 
-import Chart from 'chart.js';
 import { Store } from '@ngrx/store';
 import { AppState } from '@store';
 import * as scenesStore from '@store/scenes';
 import * as queueStore from '@store/queue';
 
 import { SubSink } from 'subsink';
-import { ChartService, ScenesService } from '@services';
+import { ScenesService } from '@services';
 import { criticalBaselineFor, CMRProduct } from '@models';
-
+import * as d3 from 'd3';
 export enum ChartDatasets {
   MASTER = 0,
   SELECTED = 1,
   DOWNLOADS = 2,
-  WITHIN_BASELINE = 3,
   PRODUCTS = 4,
   MIN_CRITICAL = 5,
   MAX_CRITICAL = 6
+}
+interface Point {
+  x: number;
+  y: number;
+  id?: string;
 }
 
 @Component({
@@ -31,22 +34,42 @@ export enum ChartDatasets {
 export class BaselineChartComponent implements OnInit, OnDestroy {
   @ViewChild('baselineChart', { static: true }) baselineChart: ElementRef;
 
-  private chart: Chart;
   private criticalBaseline: number;
-  private hoveredProductId;
   private isFirstLoad = true;
   private offsets = { temporal: 0, perpendicular: 0 };
   private subs = new SubSink();
 
+  private data = [[], [], [], [], [], [], [], []];
+  private svg;
+
+  private margin = { top: 10, right: 30, bottom: 60, left: 60 };
+  private width = 840 - this.margin.left - this.margin.right;
+  private height = 350 - this.margin.top - this.margin.bottom;
+  private dotsContainer;
+  private criticalBoxContainer;
+  private zoom;
+  private zoomBox;
+  private currentTransform;
+
+  private x;
+  private xAxis;
+  private y;
+  private yAxis;
+
+  private xExtent;
+  private yExtent;
+
+  private tooltip;
+
+  private clipContainer;
   constructor(
     private store$: Store<AppState>,
-    private chartService: ChartService,
     private scenesService: ScenesService,
   ) { }
 
   ngOnInit(): void {
-    this.initChart();
 
+    this.createSVG();
     const products$ = this.scenesService.scenes$().pipe(
       tap(products => products.map(
         product => this.criticalBaseline = criticalBaselineFor(product)
@@ -61,9 +84,9 @@ export class BaselineChartComponent implements OnInit, OnDestroy {
     );
 
     this.subs.add(
-      combineLatest(
+      combineLatest([
         products$,
-        this.store$.select(scenesStore.getSelectedScene)
+        this.store$.select(scenesStore.getSelectedScene)]
       ).subscribe(
         ([points, selected]) => {
           const extrema = this.determineMinMax(points);
@@ -77,20 +100,24 @@ export class BaselineChartComponent implements OnInit, OnDestroy {
           this.setDataset(ChartDatasets.PRODUCTS, points);
           this.setDataset(ChartDatasets.MIN_CRITICAL, minDataset);
           this.setDataset(ChartDatasets.MAX_CRITICAL, maxDataset);
-
           if (this.isFirstLoad) {
             this.updateScales(extrema);
+            this.criticalBoxContainer
+              .attr('x', this.x(this.data[ChartDatasets.MIN_CRITICAL][0].x))
+              .attr('y', this.y(this.data[ChartDatasets.MAX_CRITICAL][1].y))
+              .attr('width', this.x(this.data[ChartDatasets.MAX_CRITICAL][1].x) - this.x(this.data[ChartDatasets.MIN_CRITICAL][0].x))
+              .attr('height', this.y(this.data[ChartDatasets.MIN_CRITICAL][0].y) - this.y(this.data[ChartDatasets.MAX_CRITICAL][1].y));
             this.isFirstLoad = false;
           }
+          this.updateChart();
 
-          this.chart.update();
         })
     );
 
     this.subs.add(
-      combineLatest(
+      combineLatest([
         this.store$.select(queueStore.getQueuedProductIds),
-        this.store$.select(scenesStore.getAllProducts),
+        this.store$.select(scenesStore.getAllProducts)]
       ).pipe(
         map(([queueIds, products]) => {
           const ids = new Set(queueIds);
@@ -101,15 +128,14 @@ export class BaselineChartComponent implements OnInit, OnDestroy {
       ).subscribe(
         points => {
           this.setDataset(ChartDatasets.DOWNLOADS, points);
-          this.chart.update();
         }
       )
     );
 
     this.subs.add(
-      combineLatest(
+      combineLatest([
         this.store$.select(scenesStore.getMasterName),
-        this.store$.select(scenesStore.getAllProducts),
+        this.store$.select(scenesStore.getAllProducts)]
       ).pipe(
         map(([masterName, products]) => products
           .filter(product => product.name === masterName)
@@ -119,13 +145,157 @@ export class BaselineChartComponent implements OnInit, OnDestroy {
         map(this.productToPoint)
       ).subscribe(point => {
         this.setDataset(ChartDatasets.MASTER, [point]);
-        this.chart.update();
       })
     );
   }
+  public onResized() {
+    this.createSVG();
+  }
+  private createSVG() {
+    if (this.svg) {
+      d3.selectAll('#baseline-chart > svg').remove();
+      d3.selectAll('.tooltip').remove();
+    }
+    this.height = this.baselineChart.nativeElement.offsetHeight - this.margin.top - this.margin.bottom;
+    this.width = this.baselineChart.nativeElement.offsetWidth - this.margin.left - this.margin.right;
+    this.svg = d3.select(this.baselineChart.nativeElement).append('svg')
+      .attr('width', this.width + this.margin.left + this.margin.right)
+      .attr('height', this.height + this.margin.top + this.margin.bottom)
+      .append('g')
+      .attr('transform', `translate(${this.margin.left}, ${this.margin.top})`);
+    this.drawChart();
+  }
+  private drawChart() {
+    this.currentTransform = undefined;
+    this.clipContainer = this.svg.append('g')
+      .attr('clip-path', 'url(#clip)');
+    this.criticalBoxContainer = this.clipContainer.append('g').append('rect')
+      .attr('fill', '#f2f2f2');
+    this.zoomBox = this.clipContainer.append('rect')
+      .attr('width', this.width)
+      .attr('height', this.height)
+      .attr('cursor', 'pointer')
+      .style('fill', 'transparent')
+      .style('pointer-events', 'all');
+    this.x = d3.scaleLinear()
+    .domain(this.xExtent ?? [1, 100])
+      .range([0, this.width]);
+    this.xAxis = this.svg.append('g')
+      .attr('transform', `translate(0, ${this.height})`);
+    this.y = d3.scaleLinear()
+    .domain(this.yExtent ?? [1, 100])
+      .range([this.height, 0]);
+    this.yAxis = this.svg.append('g');
 
+
+    this.svg.append('text').attr('transform', `translate(${this.width / 2}, ${this.height + this.margin.bottom - 20})`).style('text-anchor', 'middle').text('Temporal (days)');
+    this.svg.append('text').attr('transform', `rotate(-90)`).attr('y', -this.margin.left + 20).attr('x', -this.height / 2).style('text-anchor', 'middle').text('Perpendicular (m)');
+
+    this.tooltip = d3.select('body').append('div')
+      .attr('class', 'tooltip')
+      .style('opacity', 0);
+
+    this.dotsContainer = this.clipContainer.append('g');
+    this.updateCircles();
+    this.zoom = d3.zoom()
+      .extent([[0, 0], [this.width, this.height]])
+      .on('zoom', (eve: d3.D3ZoomEvent<SVGSVGElement, unknown>) => {
+        this.currentTransform = eve.transform;
+        this.updateChart();
+      });
+
+
+    this.zoomBox.call(this.zoom);
+
+    this.svg.append('defs').append('SVG:clipPath')
+      .attr('id', 'clip')
+      .append('SVG:rect')
+      .attr('width', this.width)
+      .attr('height', this.height)
+      .attr('x', 0)
+      .attr('y', 0);
+      if (this.data[ChartDatasets.MIN_CRITICAL].length > 0) {
+        this.updateChart();
+      }
+  }
+  private updateChart() {
+    const newX = this.currentTransform?.rescaleX(this.x) ?? this.x;
+    const newY = this.currentTransform?.rescaleY(this.y) ?? this.y;
+    const smallChart = this.width > 400;
+    this.xAxis.call(
+      d3.axisBottom(newX)
+        .tickSize(-this.height)
+        .ticks(smallChart ? 10 : 5, 's')
+    );
+    this.yAxis.call(
+      d3.axisLeft(newY)
+        .tickSize(-this.width)
+        .ticks(smallChart ? 10 : 5, 's')
+    );
+
+    this.dotsContainer.selectAll('circle').data(this.data[ChartDatasets.PRODUCTS]).join('circle')
+      .attr('cx', d => newX(d.x))
+      .attr('cy', d => newY(d.y));
+
+    this.criticalBoxContainer.attr('x', newX(this.data[ChartDatasets.MIN_CRITICAL][0].x))
+      .attr('y', newY(this.data[ChartDatasets.MAX_CRITICAL][1].y))
+      .attr('width', newX(this.data[ChartDatasets.MAX_CRITICAL][1].x) - newX(this.data[ChartDatasets.MIN_CRITICAL][0].x))
+      .attr('height', newY(this.data[ChartDatasets.MIN_CRITICAL][0].y) - newY(this.data[ChartDatasets.MAX_CRITICAL][1].y));
+  }
+  private updateCircles() {
+    const self = this;
+
+    const transformedY = this.currentTransform?.rescaleY(this.y) ?? this.y;
+    const transformedX = this.currentTransform?.rescaleX(this.x) ?? this.x;
+    this.dotsContainer.selectAll('circle').data(this.data[ChartDatasets.PRODUCTS]).join('circle')
+      .attr('cx', d => transformedX(d.x))
+      .attr('cy', d => transformedY(d.y))
+      .attr('r', (d) => {
+        if (d.id === this.data[ChartDatasets.SELECTED][0].id) {
+          return 10;
+        }
+        return 5;
+      })
+      .attr('fill', function (d) {
+        if (self.data[ChartDatasets.MASTER].length > 0 && self.data[ChartDatasets.MASTER][0]?.id === d.id) {
+          return 'black';
+        } else if (self.data[ChartDatasets.SELECTED].length > 0 && self.data[ChartDatasets.SELECTED][0]?.id === d.id) {
+          return '#ff0000';
+        } else if (self.data[ChartDatasets.DOWNLOADS].some(p => p.id === d.id)) {
+          return '#215c8b';
+        } else {
+          return '#808080';
+        }
+      })
+      .on('mouseover', function (event, d: Point) {
+        self.tooltip
+          .style('opacity', .9);
+        self.tooltip.html(`${d.x} days, ${d.y} m`)
+          .style('left', `${event.pageX + 10}px`)
+          .style('top', `${event.pageY - 20}px`);
+        d3.select(this).attr('r', 10);
+      })
+      .on('mouseout', function (_event, d) {
+        self.tooltip.transition()
+          .duration(500)
+          .style('opacity', 0);
+        if (d.id === self.data[ChartDatasets.SELECTED][0].id) {
+          d3.select(this).attr('r', 10);
+        } else {
+          d3.select(this).attr('r', 5);
+        }
+      })
+      .on('click', function (_event, d: Point) {
+        const action = new scenesStore.SetSelectedScene(d.id);
+        self.store$.dispatch(action);
+      });
+  }
   private setDataset(dataset: ChartDatasets, data) {
-    this.chart.data.datasets[dataset].data = data;
+    this.data[dataset] = data;
+    if (dataset === ChartDatasets.PRODUCTS || dataset === ChartDatasets.DOWNLOADS) {
+      this.updateCircles();
+    }
+    // this.updateChart();
   }
 
   private productToPoint = (product: CMRProduct) => {
@@ -141,19 +311,19 @@ export class BaselineChartComponent implements OnInit, OnDestroy {
     const buffer = (max.x - min.x) * .25;
 
     const minDataset = [
-      {x: min.x - buffer - 10, y: -this.criticalBaseline},
-      {x: max.x + buffer + 10, y: -this.criticalBaseline}
+      { x: min.x - buffer - 10, y: -this.criticalBaseline },
+      { x: max.x + buffer + 10, y: -this.criticalBaseline }
     ];
 
     const maxDataset = [
-      {x: min.x - buffer - 10, y: this.criticalBaseline},
-      {x: max.x + buffer + 10, y: this.criticalBaseline}
+      { x: min.x - buffer - 10, y: this.criticalBaseline },
+      { x: max.x + buffer + 10, y: this.criticalBaseline }
     ];
 
     return { minDataset, maxDataset };
   }
 
-  private determineMinMax(points: {x: number, y: number}[]) {
+  private determineMinMax(points: { x: number, y: number }[]) {
     const min = { x: 0, y: 0 };
     const max = { x: 0, y: 0 };
 
@@ -175,35 +345,25 @@ export class BaselineChartComponent implements OnInit, OnDestroy {
     return { min, max };
   }
 
-  private updateScales(extrema) {
+  private updateScales(extrema: { min: { x: number, y: number }, max: { x: number, y: number } }) {
     const { min, max } = extrema;
-    const xBuffer = (max.x - min.x) * .25;
-    const yBuffer = (max.y - min.y) * .25;
-
-    this.chart.options.scales.xAxes[0].ticks.min = Math.floor((min.x - xBuffer) / 100) * 100;
-    this.chart.options.scales.xAxes[0].ticks.max = Math.ceil((max.x + xBuffer) / 100) * 100;
-
-    this.chart.options.scales.yAxes[0].ticks.min = Math.floor((min.y - yBuffer) / 100) * 100;
-    this.chart.options.scales.yAxes[0].ticks.max = Math.ceil((max.y + yBuffer) / 100) * 100;
-  }
-
-  private initChart() {
-    this.chart = this.chartService.makeChart(
-      this.baselineChart.nativeElement,
-      this.setHoveredItem,
-      this.onSelectHoveredScene
+    const xBuffer = Math.floor((max.x - min.x) * .25);
+    const yBuffer = Math.floor((max.y - min.y) * .25);
+    this.xExtent = [min.x - xBuffer, max.x + xBuffer];
+    this.x = d3.scaleLinear()
+      .domain(this.xExtent)
+      .range([0, this.width]);
+      this.yExtent = [min.y - yBuffer, max.y + yBuffer];
+    this.y = d3.scaleLinear().domain(this.yExtent).range([this.height, 0]);
+    this.xAxis.call(
+      d3.axisBottom(this.x)
+        .tickSize(-this.height)
     );
-  }
-
-  private setHoveredItem = (tooltip, data) => {
-    const dataset = data.datasets[tooltip.datasetIndex].data;
-
-    this.hoveredProductId = dataset[tooltip.index].id;
-  }
-
-  private onSelectHoveredScene = () => {
-    const action = new scenesStore.SetSelectedScene(this.hoveredProductId);
-    this.store$.dispatch(action);
+    this.yAxis.call(
+      d3.axisLeft(this.y)
+        .tickSize(-this.width)
+    );
+    this.updateChart();
   }
 
   ngOnDestroy() {
