@@ -6,7 +6,10 @@ import { MatSliderModule } from '@angular/material/slider';
 import { MatSharedModule } from '@shared';
 
 import { SeasonSelectorComponent } from './season-selector.component';
-
+import { CircleSliderModule} from '@components/shared/selectors/circle-slider/circle-slider.module';
+import { DateRangeModule } from '../date-range/date-range.module';
+import { MatButtonModule } from '@angular/material/button';
+import { PipesModule } from '@pipes';
 @NgModule({
   declarations: [SeasonSelectorComponent],
   imports: [
@@ -16,6 +19,10 @@ import { SeasonSelectorComponent } from './season-selector.component';
     MatSharedModule,
     MatSlideToggleModule,
     MatSliderModule,
+    CircleSliderModule,
+    DateRangeModule,
+    MatButtonModule,
+    PipesModule
   ],
   exports: [SeasonSelectorComponent]
 })
