@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { ReadableSizeFromBytesPipe } from './readable-size-from-bytes.pipe';
-import { ShortDatePipe, ShortDateTimePipe, FullDatePipe } from './short-date.pipe';
+import { ShortDatePipe, ShortDateTimePipe, FullDatePipe, ShortDateSeasonPipe } from './short-date.pipe';
 import { JoinPipe } from './join.pipe';
 import { BaselineFilterPipe, SBASFilterPipe, GeographicFilterPipe, ListFilterPipe } from './filter-type.pipe';
 import { HTMLInputValuePipe } from './html-input-value.pipe';
@@ -13,6 +13,7 @@ import { FilterExtensionPipe } from './filter-extension.pipe';
     ReadableSizeFromBytesPipe,
     ShortDatePipe,
     ShortDateTimePipe,
+    ShortDateSeasonPipe,
     FullDatePipe,
     JoinPipe,
     BaselineFilterPipe,
@@ -32,6 +33,7 @@ import { FilterExtensionPipe } from './filter-extension.pipe';
     ReadableSizeFromBytesPipe,
     ShortDatePipe,
     ShortDateTimePipe,
+    ShortDateSeasonPipe,
     FullDatePipe,
     JoinPipe,
     BaselineFilterPipe,
